ServerlessRAG/requirements.txt
phidata 
psycopg2 
sqlalchemy 
pgvector 
openai 
pypdf 
groq